requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
blinker==1.9.0
boto3==1.28.20
botocore==1.31.20
Brotli==1.1.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.1.0
click==8.1.7
cloudinary==1.41.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==2024.10
crispy-tailwind==1.0.3
cryptography==43.0.3
cssselect2==0.7.0
daphne==4.1.2
defusedxml==0.7.1
dj-database-url==2.3.0
dj-rest-auth==3.0.0
Django==5.1.2
django-allauth==0.54.0
django-appconf==1.0.6
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-cockroachdb==5.1
django-cors-headers==3.14.0
django-crispy-forms==2.3
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.11.2
django-filter==24.3
django-htmx==1.19.0
django-js-asset==2.2.0
django-markdownx==4.0.7
django-model-utils==5.0.0
django-nine==0.2.7
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-select2==8.2.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dnspython==2.7.0
drf-yasg==1.21.5
elastic-transport==8.4.1
elasticsearch==8.10.0
elasticsearch-dsl==8.9.0
eval_type_backport==0.2.0
facebook-sdk==3.1.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.54.1
fsspec==2024.10.0
future==0.18.3
gcloud==0.18.3
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.1
hyperlink==21.0.0
idna==3.4
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jwcrypto==1.5.6
langchain-core==0.3.12
langchain-mistralai==0.2.0
langsmith==0.1.129
lxml==5.3.0
Markdown==3.4.4
MarkupSafe==2.1.2
mistralai==1.1.0
mypy-extensions==1.0.0
oauth2client==4.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
orjson==3.10.10
packaging==24.1
pillow==10.2.0
protobuf==4.24.0
psycopg2-binary==2.9.10
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydyf==0.11.0
PyJWT==1.7.1
pymongo==4.10.1
pyOpenSSL==24.2.1
pyparsing==3.1.1
PyPDF2==3.0.1
pyphen==0.16.0
Pyrebase4==4.8.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-jwt==4.1.0
python-twitter==3.5
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
ruamel.yaml==0.17.21
s3transfer==0.6.1
service-identity==24.1.0
setuptools==75.2.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tenacity==9.0.0
tinycss2==1.4.0
tokenizers==0.20.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.7.0
zope.interface==7.1.1
zopfli==0.2.3.post1
